minor clean up
