+++ packages/node-sdk/runtime/src/mod.ts
@@ -1,25 +1,13 @@
-import { ChainBlock, genSyncProtocols } from "@paima/sync";
+import { type ChainBlock, genSyncProtocols } from "@paima/sync";
 import { getConnection } from "@paima/db";
 import { startMerge, startSync } from "@paima/sync";
 import type { SyncProtocolWithNetwork } from "@paima/config";
 import type { AppEvents } from "@paima/sm";
 import { ComponentNames, log, SeverityNumber } from "@paima/log";
-import {
-  call,
-  createChannel,
-  each,
-  type Operation,
-  spawn,
-  until,
-} from "effection";
+import { createChannel, each, type Operation, spawn } from "effection";
 import { initTelemetry } from "./telemetry.ts";
-import type { Pool } from "pg";
-import {
-  type BaseStfInput,
-  type BaseStfOutput,
-  primitiveTransitionFunction,
-} from "@paima/sm";
-import { PreparedQuery } from "npm:@pgtyped/runtime@2.4.2";
+import type { BaseStfInput, BaseStfOutput } from "@paima/sm";
+import { processFinalizedBlock } from "./process-blocks.ts";
 // import { gameStateTransitionRouter } from "@example/state-transition";
 // TODO: figure out how to setup env vars instead of relying on defaults
 const poolConfig = {
@@ -43,16 +31,7 @@ export function* start(
   ) => Promise<BaseStfOutput<AppEvents>>,
 ): Operation<void> {
   const dbConn = getConnection(poolConfig);
-  // TODO: this should be a db transaction that closes right afterwards
   const syncProtocols = yield* genSyncProtocols(dbConn, syncInfo);
-
-  // yield* call(newAddress.run({
-  //   address: "asdf",
-  // }, DBConn));
-  // const foo = yield* call(getAddressFromAddress.run({
-  //   address: "asdf",
-  // }, DBConn));
-  // console.log(foo);
 
   log.remote(
     ComponentNames.PAIMA_RUNTIME,
@@ -70,69 +49,5 @@ export function* start(
   for (const value of yield* each(finalizedBlockStream)) {
     yield* processFinalizedBlock(value, gameStateTransitionRouter, dbConn);
     yield* each.next();
-  }
-}
-
-function* executeGeneratorStepByStep<T, R>(
-  generator: Generator<T, R, unknown>,
-  dbConn: Pool,
-): Operation<R> {
-  let result = generator.next();
-
-  while (!result.done) {
-    // We resolve the generators promises here.
-    // Generators cannot execute promises.
-    // The PaimaL2 returns the state machine promise to resolve.
-    const operations: any[] = [];
-    const isPromise = result.value &&
-      typeof result.value === "object" &&
-      "type" in result.value;
-    if (
-      isPromise &&
-      (result.value as any).type === "promise"
-    ) {
-      const promiseResult = yield* until((result.value as any).promise);
-      const stateMachineQuery = (promiseResult as any).stateTransitions;
-      for (const [queryIR, params] of stateMachineQuery) {
-        const queryResult = yield* call(() => queryIR.run(params, dbConn));
-        operations.push(queryResult);
-      }
-    } else if (isPromise && (result.value as any).type === "nounce") {
-      // This operation has to persist
-      const [query, params] = (result.value as any).promise as [any, any];
-      const queryResult = yield* call(() => query.run(params, dbConn));
-      operations.push(queryResult);
-    } else if (result.value && Array.isArray(result.value)) {
-      const [queryIR, params] = result.value as [any, any];
-      const query = new PreparedQuery(queryIR);
-      const queryResult = yield* call(() => query.run(params, dbConn));
-      operations.push(queryResult);
-    }
-    result = generator.next(operations.flat());
-  }
-  return result.value;
-}
-
-// Where shoud we move this? Before emitting finalizedBlockStream?
-function* processFinalizedBlock(
-  value: ChainBlock,
-  gameStateTransitionRouter: (
-    blockHeight: number,
-    input: BaseStfInput,
-  ) => Promise<BaseStfOutput<AppEvents>>,
-  dbConn: Pool,
-): Operation<void> {
-  // TODO for this example process only evm primitives
-  if (
-    value.primitives.length > 0 &&
-    value.primitives[0].source !== "parallelUtxoRpc"
-  ) {
-    for (const primitive of value.primitives) {
-      const generator = primitiveTransitionFunction(
-        primitive as any,
-        gameStateTransitionRouter,
-      );
-      yield* executeGeneratorStepByStep(generator, dbConn);
-    }
   }
 }
